requirements.txt
anthropic>=0.20.0
python-dotenv>=1.0.0
langgraph>=0.0.40
langchain>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
streamlit>=1.30.0
requests>=2.31.0
